WATAWTAWTAWTAWTAWTAWTAWTWT it moves but has an exception

+++ Models/Location.h
@@ -22,7 +22,7 @@ public:
 	}
 	;
 
-	inline unsigned getYaw() {
+	inline double getYaw() {
 		return m_Yaw;
 	}
 	;
